apps/backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.13
anthropic>=0.39.0
python-multipart==0.0.9
aiosqlite==0.19.0
websockets==12.0
python-dotenv==1.0.0
httpx==0.27.0
